+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `Ellipse`
 - `EllipticalArc`
 - `IndexedPolycurve`
+- `IHasArcLength`
 - `Grid1d.GetCellDomains`
 - `Message.Info`
 - `Message.Error`
@@ -27,6 +28,7 @@
 
 - `Polyline` now inherits from `BoundedCurve`.
 - `Polyline` is now parameterized 0->length.
+- `Polyline` now implements the `IHasArcLength` interface
 - `Arc` now inherits from `TrimmedCurve<Circle>`.
 - `Arc` is now parameterized 0->2Pi
 - `Line` now inherits from `TrimmedCurve<InfiniteLine>`.